refactor: :recycle: make list into a seperate component
